Add timestamps to session. Simplify session by flattening structure.
We chose to simplify the structure of a session by removing the bacon
and user object and instead store the information directly in the
session object. We did this because we think the original nested
structure was too complex for our use case and the benefit of having
nested objects (readability) did not make up for the disadvantages.